used in combination with silent vsyscall switch.
This file is used for league of legends to reset the vsyscall32 value back to 1 from 0 after the game is closed. Set it as after-launch script in lutris (tick show extendend options -> system settings).

To use this in silent mode you need to:
sudo visudo

and add these two lines at the end: (it disables the password check for those two executions when executed by your user)

$USER ALL = NOPASSWD: /sbin/sysctl -w abi.vsyscall32=0
$USER ALL = NOPASSWD: /sbin/sysctl -w abi.vsyscall32=1

Note: Exchange "$USER" with your linux user name.